Move Network-related events to its own file
